Unified both auto files (womon and wmon) into one with an option
