+++ benchmarks/perception/a1_perception_2nodes/launch/trace_a1_perception_2nodes_auto.launch.py
@@ -31,15 +31,16 @@ from ros2_benchmark import ROS2BenchmarkConfig, ROS2BenchmarkTest
 
 IMAGE_RESOLUTION = ImageResolution.HD
 ROSBAG_PATH = '/home/amf/benchmark_ws/src/rosbags/perception/image' # NOTE: hardcoded, modify accordingly
-SESSION_NAME = 'a5_resize_auto_womon'
+SESSION_NAME = 'a1_perception_2nodes_auto_wmon'
 OPTION = 'without_monitor_node'
+
 
 def launch_setup(container_prefix, container_sigterm_timeout):
     """Generate launch description for benchmarking image_proc RectifyNode."""
 
     data_loader_node = ComposableNode(
         name='DataLoaderNode',
-        namespace=TestResizeNode.generate_namespace(),
+        namespace=TestRectifyNode.generate_namespace(),
         package='ros2_benchmark',
         plugin='ros2_benchmark::DataLoaderNode',
         remappings=[('camera/image_raw', 'data_loader/image'),
@@ -48,7 +49,7 @@ def launch_setup(container_prefix, container_sigterm_timeout):
 
     playback_node = ComposableNode(
         name='PlaybackNode',
-        namespace=TestResizeNode.generate_namespace(),
+        namespace=TestRectifyNode.generate_namespace(),
         package='ros2_benchmark',
         plugin='ros2_benchmark::PlaybackNode',
         parameters=[{
@@ -74,15 +75,25 @@ def launch_setup(container_prefix, container_sigterm_timeout):
         extra_arguments=[{'use_intra_process_comms': True}],
     )
 
+    rectify_node = ComposableNode(
+        name='RectifyNode',
+        namespace="robotperf/benchmark",
+        package='image_proc',
+        plugin='image_proc::RectifyNode',
+        remappings=[
+            ("image", "/robotperf/input"),
+            ("camera_info", "/r2b/camera_info"),
+        ],
+    )
 
     resize_node = ComposableNode(
+        name="ResizeNode",
         namespace="robotperf/benchmark",
         package="image_proc",
         plugin="image_proc::ResizeNode",
-        name="resize_node",
         remappings=[
             ("camera_info", "/r2b/camera_info"),
-            ("image", "/robotperf/input"),
+            ("image", "/robotperf/benchmark/image_rect"),
             ("resize", "/robotperf/benchmark/resize"),
         ],
         parameters=[
@@ -106,33 +117,60 @@ def launch_setup(container_prefix, container_sigterm_timeout):
         extra_arguments=[{'use_intra_process_comms': True}],
     )
 
-    composable_node_container = ComposableNodeContainer(
-        name='container',
-        namespace=TestResizeNode.generate_namespace(),
-        package='rclcpp_components',
-        executable='component_container_mt',
-        prefix=container_prefix,
-        sigterm_timeout=container_sigterm_timeout,
-        composable_node_descriptions=[
+    monitor_node = ComposableNode(
+        name='MonitorNode',
+        namespace=TestRectifyNode.generate_namespace(),
+        package='ros2_benchmark',
+        plugin='ros2_benchmark::MonitorNode',
+        parameters=[{
+            'monitor_data_format': 'sensor_msgs/msg/Image',
+        }],
+        remappings=[
+            ('output', '/robotperf/benchmark/resize')],
+    )
+
+    if OPTION == 'with_monitor_node':
+        composable_node_descriptions_option=[
             data_loader_node,
             # prep_resize_node,
             playback_node,
             input_node,
+            rectify_node,
             resize_node,
-            output_node            
-        ],
+            output_node,
+            monitor_node  
+        ]
+    else:
+        composable_node_descriptions_option=[
+            data_loader_node,
+            # prep_resize_node,
+            playback_node,
+            input_node,
+            rectify_node,
+            resize_node,
+            output_node,  
+        ]
+
+    composable_node_container = ComposableNodeContainer(
+        name='container',
+        namespace=TestRectifyNode.generate_namespace(),
+        package='rclcpp_components',
+        executable='component_container_mt',
+        prefix=container_prefix,
+        sigterm_timeout=container_sigterm_timeout,
+        composable_node_descriptions=composable_node_descriptions_option,
         output='screen'
     )
 
     return [composable_node_container]
 
 
-class TestResizeNode(ROS2BenchmarkTest):
-    """Performance test for image_proc RectifyNode."""
+class TestRectifyNode(ROS2BenchmarkTest):
+    """Performance test for image_proc RectifyNode and ResizeNode."""
 
     # Custom configurations
     config = ROS2BenchmarkConfig(
-        benchmark_name='image_proc::RectifyNode Benchmark',
+        benchmark_name='image_proc::RectifyNodeResizeNode Benchmark',
         input_data_path=ROSBAG_PATH,
         # Upper and lower bounds of peak throughput search window
         publisher_upper_frequency=30.0,
@@ -167,5 +205,7 @@ class TestResizeNode(ROS2BenchmarkTest):
 
 
 def generate_test_description():
-    return TestResizeNode.generate_test_description_with_nsys(launch_setup)
+    return TestRectifyNode.generate_test_description_with_nsys(launch_setup)
+
+
